update readme with game link
I forgot to add the link to Heroku on the submission of this project. There is still a week left for the submission deadline and I'm not sure if I have to resubmit the whole thing or if I can just add the small change to your fork here.

Please advise. Thanks!